Split MIDI editor into multiple files
